util クラスを追加して3つのクラスを移動させました。コメントを追加しました。
