+++ README.md
@@ -1,17 +1,64 @@
-# BingoGame
-5月研修用ビンゴゲーム
-
 ビンゴゲーム概要
-* 現在はプレイヤー2人までを想定。
-*ビンゴカードに振られる数字は、実際のビンゴカードと同じく左から、
-	1列目：1～15
-	2列目：16～30
-	3列目：31～45(FreeSpaceあり）
-	4列目：46～60
-	5列目：61～75
-* 縦・横・ななめのどれかで「●」を5つそろえたら終了。
-* 誰かが一人でもビンゴを出した場合には即終了となる。
+---------------------------------------
+
+* マネージャとプレイヤーの対戦で行います。
+
+* マネージャはプレイヤー名が Manager です。
+
+* プレイヤーはゲーム開始時に名前を入力して決定します。
+
+* マネージャは、ビンゴボールを1つずつ取り出し、プレイヤーに通知します。
+
+* プレイヤーは通知されたボールの番号を判別して自分のビンゴカードで一致する番号があると xx に置き換えます。
+
+* ビンゴカードに振られる数字は、以下となります。
+
+	1 列目：1～15
+	2 列目：16～30
+	3 列目：31～45(真ん中は FreeSpace で xx が設定されています）
+	4 列目：46～60
+	5 列目：61～75
+
+* マネージャ、またはプレイヤーのどちらかが縦・横・ななめのどれかで xx を 5 つそろえたら bingo となり終了です。
+
 * ゲーム回数は10回までと設定
 	→回数制限を設けないと30回やってもビンゴにならない可能性があるため。
+
 * 最初のボール排出前に、まず10玉排出する。
 	→ゲーム回数を減らしているため
+
+
+com.foo.training.bingo.base パッケージ
+---------------------------------------
+プレイヤーの基底クラス(BaseBingoPlayer)、マネージャインタフェイス(BingoManager) が配置されています。
+
+
+com.foo.training.bingo.main パッケージ
+---------------------------------------
+ゲームを行うためのメインクラス(BGame) のみ配置されています。
+
+
+com.foo.training.bingo.player パッケージ
+---------------------------------------
+プレイヤー具象クラス(Player)とプレイヤー具象クラスにマネージャを実装したクラス(Manager) が配置されています。
+
+
+com.foo.training.bingo.util パッケージ
+---------------------------------------
+定数保持クラス(BingoConstant)、カード生成クラス(BingoCard)、ボール管理クラス(BingoBall) が配置されています。
+
+
+com.foo.training.bingo.test パッケージ
+---------------------------------------
+JUnit テストクラスが配置されています。
+
+このパッケージ配下で、作業をしてもらうのは以下です。
+
+* BaseBingoPlayerTest クラスにおいて、com.foo.training.bingo.player パッケージ配下で実装した Player クラスのリーチ、ビンゴ時の叫び声を上書きしてください
+
+* data パッケージ配下の HorizontalReachTestValue, VerticalReachTestValue の各クラスでテストケースのデータが不足しているので、補ってください。
+
+
+上記 2 点以外のコードは修正しないでください。
+
+パッケージ配下のテストクラスを実行して全てのテストが正常に完了することを確認してください。
