+++ src/com/foo/training/bingo/main/BGame.java
@@ -2,6 +2,11 @@ package com.foo.training.bingo.main;
 
 import com.foo.training.bingo.player.Manager;
 
+/**
+ * Bingo ゲームを進めるメインクラス
+ * @author
+ *
+ */
 public class BGame {
 
 	public static void main(String[] args) {
@@ -12,7 +17,6 @@ public class BGame {
 //		BaseBingoPlayer player = mngr.getPlayer(playerName);
 //		player.showMyCard();
 		mngr.progressGame();
-		mngr.showAllPlayerCard();
 	}
 
 }
