+++ src/com/foo/training/bingo/player/Player.java
@@ -2,6 +2,11 @@ package com.foo.training.bingo.player;
 
 import com.foo.training.bingo.base.BaseBingoPlayer;
 
+/**
+ * ビンゴプレイヤークラス
+ * @author
+ *
+ */
 public class Player extends BaseBingoPlayer {
 
 	public Player(String name, String[][] card) {
