+++ src/com/foo/training/bingo/test/BingoPlayerVBingoTest.java
@@ -14,6 +14,7 @@ import com.foo.training.bingo.test.data.VerticalBingoTestValue;
 
 /**
  * BaseBingoPlayer のテストクラス
+ * 垂直 Bingo が正しく判別されることを確認します。
  * @author
  *
  */
@@ -30,7 +31,7 @@ public class BingoPlayerVBingoTest extends BaseBingoPlayerTest {
 			};
 
 	/**
-	 * 垂直でBingoの場合に正しく叫ぶか否かをテストします。
+	 * 垂直で Bingo の場合に正しく叫ぶか否かをテストします。
 	 */
 	@Theory
 	public void testVerticalBingoValues(String[][] vbvalue) {
