+++ src/com/foo/training/bingo/test/base/BaseBingoPlayerTest.java
@@ -6,6 +6,11 @@ import java.io.PrintStream;
 import org.junit.After;
 import org.junit.Before;
 
+/**
+ * BingoPlayer 用テスト基底クラス
+ * @author
+ *
+ */
 public abstract class BaseBingoPlayerTest {
 	/**
 	 * TODO
