+++ src/com/foo/training/bingo/test/data/VerticalBingoTestValue.java
@@ -1,5 +1,10 @@
 package com.foo.training.bingo.test.data;
 
+/**
+ * 垂直 Bingo テスト用データクラス
+ * @author
+ *
+ */
 public class VerticalBingoTestValue {
 
 	public static String[][] getCard1() {
