+++ src/com/foo/training/bingo/util/BingoConstant.java
@@ -1,0 +1,34 @@
+package com.foo.training.bingo.util;
+
+/**
+ * 定数保持クラス
+ * @author
+ *
+ */
+public final class BingoConstant {
+
+	/**
+	 * 番号が一致した場合の代替文字列
+	 */
+	public static final String done = "xx";
+
+	/**
+	 * ビンゴの玉の最大値
+	 */
+	public static final int maxBallNumber = 75;
+
+	/**
+	 * カード要素のサイズ
+	 */
+	public static final int matrixSize = 5;
+
+	/**
+	 * 事前ゲーム実施回数
+	 */
+	public static final int preGameCount = 10;
+
+	/**
+	 * ゲーム回数
+	 */
+	public static final int gameCount = 10;
+}
